Session-3-Practical-Use-Case-Evaluations/requirements.txt
# Strands Agent SDK and instrumentation
strands-agents[otel]
strands-agents-tools
boto3
botocore
ddgs
requests
pypdf
numpy
aws-opentelemetry-distro~=0.12.1
python-dotenv